+++ CHANGELOG.md
@@ -15,3 +15,9 @@
 
 ## 0.0.5 10.Oct.2021
 - `gxscalc.mtp()` has been changed to return (None, None) when no cross points are detected.
+
+## 0.1.0 09.Aug.2023
+- Slightly changed the algorithm of `gxscalc.mtp()`.
+- Add estimated MT point in the figure of `gxscalc.mtp()`.
+- Add `with_frames` keyword argument into `gxscalc.mtp()` and `gxscalc.distance`.
+- `gxscalc.mtp()` and `gxscalc.distance` accept direct DataFrame inputs.